backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-filter>=23.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
django-cors-headers>=4.0.0

# Web Scraping Libraries
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0
selenium>=4.15.0
httpx>=0.25.0

